dreambooth/requirements.txt
accelerate==0.16.0
bitsandbytes==0.35.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.22.6
diffusers @ git+https://github.com/ShivamShrirao/diffusers@25045fde9c4281ee317703714cc0331bb97d5a65
filelock==3.9.0
huggingface-hub==0.12.0
idna==2.10
importlib-metadata==6.0.0
lit==16.0.0rc1
mypy-extensions==1.0.0
numpy==1.21.6
packaging==23.0
Pillow==7.1.2
psutil==5.4.8
pyre-extensions==0.0.23
PyYAML==6.0
regex==2022.6.2
requests==2.25.1
safetensors==0.2.8
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
triton==2.0.0a1
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.24.3
xformers==0.0.17.dev451
zipp==3.12.1
